suarez/requirements.txt
Django==6.0.1
pillow==12.1.0
python-dotenv==1.0.1
whitenoise==6.8.2
django-csp==4.0

xhtml2pdf==0.2.17
reportlab==4.4.9

gunicorn==23.0.0
psycopg2-binary==2.9.9